Cleanup into a spec/support/features module.
