+++ spec/features/add_to_slack_spec.rb
@@ -2,8 +2,6 @@ require 'rails_helper'
 
 RSpec.feature 'Add to Slack' do
   fixtures :all
-  include Rack::Test::Methods
-  let(:app) { SlashDeploy.app }
 
   scenario 'clicking the Add to Slack button' do
     stub_request(:post, 'https://slack.com/api/oauth.access')
