+++ spec/fixtures/repositories.yml
@@ -1,0 +1,3 @@
+'baxterthehacker/public-repo':
+  name: 'baxterthehacker/public-repo'
+  github_secret: secret
