+++ spec/rails_helper.rb
@@ -54,4 +54,7 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
+
+  require_relative '../spec/support/features'
+  config.include Features, type: :feature
 end
